requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
cffi==1.11.5
colorama==0.4.0
cryptography==2.3.1
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.1.1
PyMySQL==0.9.2
pytz==2018.7
six==1.11.0
wrapt==1.10.11
